requirements.txt
streamlit==1.28.1
google-genai==0.3.0
selenium==4.15.2
webdriver-manager==4.0.1
